[#23] Updated to run commands from CLI or filter down scripts.
